+++ ledger/src/test_utilities.rs
@@ -891,7 +891,7 @@ impl ProvingProvider for ProofServerProvider<'_> {
             println!("    Proving response: {} bytes", bytes.len());
             let proof: ProofVersioned = tagged_deserialize(&mut bytes.to_vec().as_slice())?;
             match proof {
-                ProofVersioned::V2(proof) => Ok(proof),
+                ProofVersioned::V2(proof) | ProofVersioned::V3(proof) => Ok(proof),
             }
         } else {
             anyhow::bail!(